Add TableRow component so each task can be marked off individually
